backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
langchain-community
sse-starlette
pydantic
python-dotenv
aiohttp
requests
